% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_board.R
\name{create_board}
\alias{create_board}
\title{Create the game board}
\usage{
create_board(size = 10, land_pct = 0.2)
}
\arguments{
\item{size}{Size of the board (default 10, max 20).}

\item{land_pct}{Percentage of land cells (default 0.2, max 0.5).}
}
\value{
A data frame with board coordinates and states.
}
\description{
Create the game board
}
