% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fleet.R
\name{create_fleet}
\alias{create_fleet}
\title{Create the fleet}
\usage{
create_fleet(board, carrier = 1, destroyer = 1, frigate = 2, patrol = 3)
}
\arguments{
\item{board}{A board created with create_board().}

\item{carrier}{Number of carriers (size 5).}

\item{destroyer}{Number of destroyers (size 4).}

\item{frigate}{Number of frigates (size 3).}

\item{patrol}{Number of patrol boats (size 2).}
}
\value{
A data frame describing the fleet.
}
\description{
Create the fleet
}
