% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_ships.R
\name{place_ships}
\alias{place_ships}
\title{Place ships on the board}
\usage{
place_ships(board, fleet)
}
\arguments{
\item{board}{A board from create_board().}

\item{fleet}{A fleet from create_fleet().}
}
\value{
Updated board with ships placed.
}
\description{
Place ships on the board
}
